backend/requirements.txt
fastapi
uvicorn
sse-starlette
pydantic
langgraph
langchain-google-genai
python-dotenv
docker
requests